add command files for wxpo json
